backend/requirements.txt
# Tradie Migration App – Python Requirements
# Install with: pip install -r requirements.txt
# Compatible with Python 3.12, 3.13, 3.14

# ── Web Framework ──────────────────────────────────────────────────────────────
fastapi>=0.115.0
uvicorn[standard]>=0.30.6

# ── Database ───────────────────────────────────────────────────────────────────
sqlalchemy[asyncio]>=2.0.35
asyncpg>=0.30.0            # Async PostgreSQL driver (0.30+ supports Python 3.13/3.14)
alembic>=1.13.3            # DB migrations
pgvector>=0.3.6            # pgvector ORM integration (for RAG embeddings)
psycopg2-binary>=2.9.9     # Sync driver (for Alembic migrations)

# ── AWS ────────────────────────────────────────────────────────────────────────
boto3>=1.35.0              # S3 presigned URLs, Bedrock (AI), Cognito, Secrets Manager

# ── Auth / Security ────────────────────────────────────────────────────────────
python-jose[cryptography]>=3.3.0   # JWT validation (Cognito tokens)
passlib[bcrypt]>=1.7.4

# ── Data Validation ────────────────────────────────────────────────────────────
pydantic>=2.10.0           # 2.10+ has pre-built wheels for Python 3.14
pydantic-settings>=2.6.0
python-multipart>=0.0.12   # File upload support

# ── Document Processing (RAG background Lambda) ────────────────────────────────
pypdf2>=3.0.1              # PDF text extraction
python-docx>=1.1.2         # Word document extraction

# ── Utilities ──────────────────────────────────────────────────────────────────
python-dotenv>=1.0.1       # Local .env support
httpx>=0.27.2              # Async HTTP client for testing

# ── Testing ────────────────────────────────────────────────────────────────────
pytest>=8.3.3
pytest-asyncio>=0.24.0